log view: fix page loading and retention
we reused the 'pending_pages' map to determine which pages are still
required, but that made problems when crossing a page boundary when
scrolling under certain circumstances. In addition the pages where not
ordered in the cache, leading to changes in the dom markup that were not
necessary and also lead to flickers and glitches.

To fix those issues, we now track the required pages separately, order
the pages after inserting, and use our components with `.key()` to help
yew with dom diffing.

Also move the removal from `pending_pages` to the top, so it always gets
removed from there.

This also factors out the 'insert_page' code, which makes it a bit more
clearer what the code does.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>