journal view: implement functionality
uses the api calls/journalctls start/end cursor to optimizie what to
load. When scrolled to the bottom, tries to load new lines in a 1000ms
interval beginning with the saved end cursor, and when scrolling to the
top, loads 500 lines from before the saved start cursor.

when scrolled somewhere in between, there is no loading.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>