rrd graphs: fix curves that contain positive and negative values
by not creating two fill paths (one for the positive and one for the
negative values), but drawing one fill path for both simultaneously

This works because the path starts and ends at the y zero coordinate,
which then can connect cleanly.

At the end of data we also have to use the previous x value, otherwise
we draw the filling to the next (undefined) point instead of straight
to the y axis.

This fixes rendering issue when the value changes from negative to
positive.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>