rrd_timeframe_selector: make RRDTimeframe more flexible
RrdMode and RrdTimeframe are now available via the wasm-compatible
proxmox-rrd-api-types crate.

The PdmClient takes these as typed parameters to its rrddata methods,
but the current yew-comp RRDTimeframe type does not provide access to
them separately, but instead combines the two (which is a bit of a
name confusion which should probably be solved in a later commit) into
an enum which serializes into a serde_json::Value::Object with the two
components stringified.

Instead, make this a struct containing both members publicly so we can
just pass them through for calls like
    pdm_client().pve_node_rrddata(&remote, &node, tf.mode, tf.timeframe)

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>