wizard: add reset_remaining_valid_pages
sometimes, a wizard page changes the state of the remaining wizard, so
we want to have the possibility to reset the valid page state for those
pages.

implement this by simply iterating over tha inserted paged in the
correct order, and start changing the valid state after the current key.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>