js-helper/http-client: set SameSite and Secure properties
`SameSite=Lax` is the future default if no SameSite property is set and
setting it explicitly removes a warning from the logs. this patch
sets it when setting the auth cookies in `HttpClientWasm` so that
other invocations of the `set_cookie` function in
`js-helper-module.js` can still specify different values as needed.

however, in the case of `clear_auth_cookie` it is added there. as the
cookie is intended to immediately expire anyway, the value does not
matter, but setting it explicitly removes the warning in the console.

`Secure` is used to send cookies only via https. this is important as
leaking authentication to all possible parties in between the user
and the server is not desirable.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>