acme/notes view: remove compiler warnings about never type fallback ()
previously the never type (`!`) would fall back to the unit type. this
is going to be phased out and become a hard error. so explicitly set
the return type to unit (`()`), which in most cases can be achieved
simply by letting the compiler infer the return type of the function
call from the return type of the function that calls it.

also see: https://github.com/rust-lang/rust/issues/123748

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>