buildsys: replace comically large h3 with h6 in package info tooltip
In a similar spirit as 3c62093 ("apt changelog: use lower h6 heading
for first line").

Add comment for rethinking this, tooltips are not ideal for such
general information, they should be mostly used for hints and form
(validity) feedback. Here it might be better to have a actual "show
package info" (action) button that opens up a window with all the
information. Could be also nice to see that for all installed
packages, but that all is a bit of a bigger change and should be done
in our ExtJS UIs too, so no need to worry about it now.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>