acme: use `as_str` instead of re- and dereferencing
this check trips up clippy fixes, as `&*` should be redundant if both
sides try to take reference. however, due to how this type is
encapsulated, that won't work and break the check. so use `as_str` to
get a reference to a string slice for with comparison with `String` is
properly implemented.