rrd graph: cope with out-of-bound selection
Avoid panicking due to unconditionally access the data0 slice, rather
use get and see if we actually got data to work with, if not then
ignore it when the data slice is empty or log a debug message
otherwise, as that case shouldn't really happen.

The diff is consisting mostly of a indentation change, so best to
checkout with git's `-w` flag to ignore white space change.

This was reported in our community forum [0].

[0]: https://forum.proxmox.com/threads/159485/

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>