HttpClient: add helpers to refresh HttpOnly cookies and remove them
the `refresh` function handles a refresh without needing a ticket, the
assumption is that the ticket is handled by an HttpOnly cookie that we
don't have access to.

the `logout` function requests a removal of the ticket by the server
as we cannot remove properly protected cookies from the client-side.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>