wizard: factor out 'can progress' logic check to WizardState
with that we have a single point where we can check if the current
form/page state should allow us to progress.

The existing logic was reversed to get a 'positive' variable
(can_progress vs next_is_disabled), so we have to adapt the used
variable names and checks a bit.

No functional change intended.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>