wizard: fix form progressing when pressing enter
commit
d5f3c95 (wizard: allow enter to be used for switching to next page)

intended the current page to be progressed with enter the same way
the next button is pressed.

The check for this was wrong though, since not only the form context
validity is relevant, also the page_lock state and previous page
validity state.

To fix this, use the WizardState's `can_progress` check we now have
instead, which is also used to enable/disable the `next` button.

Fixes: d5f3c95 (wizard: allow enter to be used for switching to next page)
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>