fix #6382: tasks: don't add duplicate tasks into the list
When loading tasks due to reaching the end of the list, the amount of
existing tasks was used as the start parameter for the next load.

New tasks are added at the beginning of the list, so if new tasks were
added in the meantime, the API call would return tasks we already had
in the list. The UPID of the tasks are getting used as keys for the
store to uniquely identify the records. Since those get also used as yew
keys for the vdom by the data table and duplicate keys are not supported
in yew (can lead to a crash) we must avoid this.

To fix this, instead of using the amount of tasks as offset, use the
last task's start time as new 'until' value. With this we know there will
be overlap, but only the tasks started the same second. We can skip them
by finding the cutoff in the newly returned data which is the index of
the UPID of the last one of the current store content.

This makes the 'start' parameter of the LoadBatch update unnecessary,
and we replace it with a 'fresh load' flag that get's reset after every
load, and only gets used by the refesh button and when filters change.
(like we used 'start == 0' before)

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>