rrd: units: simplify calculations for get_grid_unit_base
in get_grid_unit_base2, when resolving the mathematical calculations,
one can see that we always subtract 2 from the log2 to get to the
result. (with the exception of ranges smaller than 2 and very big
numbers where the floating point operation does not work properly
anymore in 64 bits, e.g. 2.0^1024 -> Infinity). This saves us the loop
and at least 2 calculations for the condition.

The second condition could never be reached, since doubling a number <
4.0 can never be > 15.0, so we can just remove it.

In get_grid_unit_base10, we can at least change the while loop to an if,
since that will only ever run once.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>