rrd: remove wrongly annotated lifetime
the data returned from get_view_data does not depend on self, but the
way the lifetimes are annotated implies this. This can be a problem when
we want to borrow self mutably, but want to keep the data returned from
this method. By simply removing the lifetime annotation from &self, we
tell the compiler that we don't have to borrow &self immutably

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>