rrd: refactor grid data computation
namely the data/time min/max, interval and ranges. This makes the main
graph module a bit smaller, since we can move the compute_min_max
functionality out.

The grid unit calculation can then be private in the units module.

Rename the variables so they are consistent (*_min, *_max, *_interval,
etc.)

The check for start/end time now can be simplified to data0.is_empty(),
because we don't need to extract the first and last time a second time.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>