rrd: introduce GraphSpace struct and use it to precalculate graph data
This contains all information about the data and the svg graph that is
needed to convert coordinates between data and svg space.

This replaces the LayoutProps struct on the graph, and provides the
necessary interfaces.

It's data is updated whenever the grid parameters are changed so that we
need to recalculate them, e.g. when the data changes, or the width of
the grid. Since this is now only done when necessary, it should be much
faster when unrelated redraws occur, e.g. when moving the mouse cursor
over the data.

We need to derive PartialEq for the Series struct to check if they have
changed.

Also change the interfaces of compute_*_path, so that they only take
a GraphSpace struct anymore instead of the min/max data and compute
functions (all that is provided by the struct).

`compute_fill_path` calculation of y0 is changed slightly to accommodate
the fact that the GraphSpace only takes values inside the min and max
data ranges, but the end result should be the same.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>