rrd: precalculate the grid line and label positions
Introduce a new `RrdGrid` struct, that calculates the correct positions
of the grid lines and label positions with the help of a `GraphSpace`
struct.

Since the calculation is now split off from the creation of the
component, the code becomes much clearer, and the code in the main graph
module becomes much smaller.

It saves only the base data, not the actual component themselves, they
are generated via helper methods on the fly.

This should speed up all operations on the graph that don't need to
update the grid lines, like moving the mouse cursor over the data.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>