http helpers: expose ticket refresh loop controls
by introducing a stop_ticket_refresh_loop and making both start and stop
public. This can be useful e.g. when such a ticket refresh loop is
unwanted. This is the case for example for the PMG quarantine interface
when using a PMGQUAR ticket, since that is valid as long as the
quarantine is valid and refreshing it simply yields the same ticket.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>