make: test all-features and all-targets
To increase the chance that one pushes changes that introduce compile
or test errors for optional features. It would be better if debcargo
would learn to do this though.

Note that the targets in "all-targets" doesn't mean compile
architecture targets like wasm or amd64 but rather cargo (output)
targets like examples, bins, lib, benches and tests.

Signe-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>