rrd: fix rrd tooltip alignment for edge of viewport
This is a semantical revert of:
 023c856 (rrd: align tooltip directly to pointer position)

We saved a bit of code by aligning the tooltip to the mouse position
directly by using align_to_xy, but this meant that the tooltip got
outside the viewport. This happens because align_to_xy does not
automatically keep the element to be aligned inside the viewport.

Instead of trying to add that functionality there, simply do what we did
before: adding an invisible element into the svg where we can align the
tooltip to.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>