auth_edit_ldap: improve form layout and placeholders
make certain field large as they may contain too much information to
fit into a small field. also switches usages of `company.net` to
`exmaple.com` in compliance with rfc2606 [1]. also hides the bind dn's
placeholder if anonymous search is enabled to not give the impression
that a default value is used for the bind dn.

[1]: https://datatracker.ietf.org/doc/html/rfc2606

Suggested-by: Christoph Hess <c.heiss@proxmox.com>
Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250922150519.399573-8-s.sterz@proxmox.com