add `node_info` helper to render a consistent view of the NodeStatus
since we want this for the pve and pbs NodeStatus, we introduce a small
enum type that can hold a reference to either and dynamically extract the
information we need.

That way we can easily use the helper for both PVE and PBS node panels.

The content itself should be the same as the one we have in the native
PVE and PBS interface, minus the subscription info.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>