node info: avoid overly low warning/critical tresholds for memory usage
This mirrors a recent change in pve-manager with commit 5b6bfa4a1
("ui: guests status: increase threshold for warning/critical memory
usage"), i.e. memory is there to be used and leveraging most of what's
configured is an OK thing to do, so it's rather bad UX to show a
warning status at already 70% of usage (the default for the zero
optimum meter was only bumped from 70% to 80% recently).

This change was done in PDM but got lost once PDM switched over to the
common implementation here.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>