http client: update to newer web-sys RequestInit interfaces
Use the new setters which are not builder style and thus we can also
loose the mutability.

Background: Web-sys deprecated the old setters due to using a naming
style that is normally used for getters. They also added actual
getters but for now with the get_ prefix to avoid a breaking change.

See wasm-bindgen commit 028dbc94a129 ("Generate getters for all WebIDL
dictionary types ") and PR [#3993] for details.

[#3993]: https://github.com/wasm-bindgen/wasm-bindgen/pull/3993

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>