token panel: improve token secret dialog layout and hide password
the dialog showing the token secret now displays it like any other
password. meaning that the actual secret is covered up at first but
can be displayed if required. this adds protections against should
surfing attacks.

the copy secret button has also been moved next to the input field and
is now an icon buton with a tool tip.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251017124646.294343-5-s.sterz@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>