fix #6787: ui: allow creating PAM users from dashboard
When tried to create a user in the PAM realm an error occurs if the user
doesn't exist on the host. To fix the issue hid the password and confirm
password fields. Now the user will only be added to the `.cfg` file and
won't be trying to update the password. The error was originating from
the invocation of `pam_chauthtok`.

Since the user is not created on the system, if tried to change the
password the same error will be shown again. Inorder to avoid that
the "change password" button will only be enabled if the user is
not a "pam" user.

Signed-off-by: Shan Shaji <s.shaji@proxmox.com>
Link: https://lore.proxmox.com/20251028133326.342164-1-s.shaji@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>