apt repositories: add 'status_only' property
If enabled, this will only show the upper status message table (e.g. 'You
get updates for ...').

While it would be better to split out the status message table into its
own little component, this is a much bigger task than just adding the
new property and a couple of if-guards in the render function. Given the
time constaints this was the better choice. Extracting the component
can always be done at a later point in time.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251127104447.162951-6-l.wagner@proxmox.com