apt repo: add subscription_url widget property
In the same spirit as the base_url but for the subscription status,
required e.g. in the PDM, where we currently mixed the subscription
status of the PDM itself with the repository status of the remote,
leading to rather confusing results.

In the midterm it might be nicer to rather adapt this to (optionally)
pass the repo and subscription status instead directly, but for now
this is definitively the simpler and less invasive stop-gap.

Use the current value as default so that we stay backward compatible.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>