apt package manager: add distinct message type to avoid spurious subscription popups
The freshly introduced subscription check works via sending a message to
the component which is then processed by the `update` function. When
implementing, it was overlooked that there is was already a selection
change listener set up to send messages on selection change, which led
to the behavior of triggering the subscription check whenever a package
was selected by the list.

This is fixed by adding a distinct Message enum type with variants for
the selection change event and the check subscription event.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251202100239.117392-1-l.wagner@proxmox.com