subscription check: use more useful function signature
having a reference of an option is very unweildly at times, e.g. when
having the inner type, it's necessary to move it into an option, even if
`subscription_is_active` does not need to own the data.

using an option of a reference makes usage much more ergonomic at times.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251202143226.3681712-2-d.csapak@proxmox.com