fix: add a `div` wrapper element around the notes content
When the <script> element is placed on the first line of the note,
the `DomParser` parses it and moves it into the <head> element. As a
result, the <body> does not contain the <script> element, and the
NoteView fails to display it. However, when the <script> element is
placed between two lines, it is correctly placed inside the <body>.

To fix this issue, a wrapper <div> element has been added around the
note string.

Signed-off-by: Shan Shaji <s.shaji@proxmox.com>
Link: https://lore.proxmox.com/20251202150109.455810-1-s.shaji@proxmox.com