wizard: prevent page reload when pressing enter in form
When pressing enter while the focus is inside a `Form` element, the
browser tries to submit the form, and the default method is to send the
data to the page via GET parameters.

This can lead to unwanted refreshes, e.g. in PDM the remote wizard would
change the url from
 https://<hostname>:8443/#/
to
 https://<hostname>:8443/?#/

which refreshed the page.

To prevent that, use `event.prevent_default()` in the onsubmit handler,
since we handle submitting the data ourselves there.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251211133223.2607590-1-d.csapak@proxmox.com