user panel: don't submit raw "expire" field
the api expects a unix timestamp here, but submitting the raw field
will submit a rfc3999 timestamp. this happens when a user has expire
set and is updated, but the expire field itself isn't changed. the ui
would then submit the raw expire field. to fix this, don't submit the
field by itself.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251215133951.210548-1-s.sterz@proxmox.com